feat. add Lotto Constructor with argument(String)
