+++ src/main/java/com/zingoworks/lotto/view/ResultView.java
@@ -23,7 +23,8 @@ public class ResultView {
         System.out.println(NEWLINE + "당첨통계" + NEWLINE + "---------");
         for (Prize prize : Prize.values()) {
             System.out.println(prize.getScore().getCountOfHit()
-                    + "개 일치" + getBonusHitPhrase(prize.getScore().isBonusHit()) + "(" +  prize.getWinningMoney() + "원) - "
+                    + "개 일치" + getBonusHitPhrase(prize.getScore().isBonusHit())
+                    + "(" +  prize.getWinningMoney() + "원) - "
                     + lottoResults.getCountOfWins(prize.getScore()) + "개");
         }
         System.out.println("총 수익률은 " + lottoResults.getNetEarningRate(purchaseAmount) + "%입니다.");
